requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-braces==1.14.0
django-cors-headers==3.7.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
Pillow==8.1.2
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
whitenoise==5.3.0
wincertstore==0.2
